+++ fs-storage/src/jni/mod.rs
@@ -1,0 +1,1 @@
+pub mod file_storage;
